backend/pyproject.toml
[project]
name = "rag-with-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-core>=1.38.0",
    "chromadb>=1.4.1",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=0.6.7",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pandas>=2.3.3",
    "pypdf2>=3.0.1",
    "pysqlite3>=0.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.2.2",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.40.0",
]
